refactor(nerva): move NervaResult and MapServiceToProtocol to pkg/brutus
Extract nerva service discovery types and protocol mapping from
cmd/brutus/mapping.go into pkg/brutus/nerva.go. Callers in runner.go
and integration_test.go now use the exported brutus.NervaResult and
brutus.MapServiceToProtocol.
